Fix kprobe to work in modern kernels
This PoC was developed to run on quite old kernels, where
ns_capable_common was not inlined and had access to a bit more
information than cap_capable in some cases. Nowadays the only
way to reliably get results is to kprobe cap_capable.

Since its kernel interface changed over time, we use a simple
check on whether arg3 is a kernel pointer. This allows the code
to remain simple whilst still working on both modern and older
kernels.